backend/requirements.txt
firebase-admin>=6.5.0
python-dotenv>=1.0.1
fastapi-utils>=0.8.0
